+++ constants/messages/messages.go
@@ -1,5 +1,5 @@
 package messages
 
 const (
-	OfflineMessage = "Error fetching submissions. Press 'r' to retry or 'q' to quit."
+	OfflineMessage = "Could not fetch submissions. [-:-:] Press 'r' to retry or 'q' to quit."
 )